menu de add cartas 100% feito

+++ include/struct_card.h
@@ -9,8 +9,8 @@ typedef struct{
     char b_name[20];
     bool b_type[4];
     bool b_trunfo;
-    int b_atk;
-    int b_hp;
+    float b_atk;
+    float b_hp;
     float b_pso;
     float b_alt;
     int b_hab;
